Add dialog destination example
