+++ compose/mpp/demo/src/commonMain/kotlin/androidx/compose/mpp/demo/FontFamilies.kt
@@ -16,18 +16,18 @@
 
 package androidx.compose.mpp.demo
 
-import androidx.compose.foundation.border
-import androidx.compose.foundation.layout.*
+import androidx.compose.foundation.layout.Arrangement
+import androidx.compose.foundation.layout.Column
+import androidx.compose.foundation.layout.fillMaxSize
+import androidx.compose.foundation.layout.padding
 import androidx.compose.foundation.rememberScrollState
 import androidx.compose.foundation.verticalScroll
 import androidx.compose.material.MaterialTheme
 import androidx.compose.material.Text
 import androidx.compose.runtime.Composable
 import androidx.compose.ui.Modifier
-import androidx.compose.ui.graphics.Color
 import androidx.compose.ui.text.font.FontFamily
 import androidx.compose.ui.unit.dp
-import androidx.compose.ui.unit.sp
 
 
 @Composable
@@ -55,7 +55,7 @@ fun FontFamilies() {
 }
 
 @Composable
-fun FontFamilyShowcase(fontFamily: FontFamily) {
+private fun FontFamilyShowcase(fontFamily: FontFamily) {
     Column {
         Text(
             text = "$fontFamily"
